validation3

+++ app/Http/Controllers/admin/ColorController.php
@@ -11,6 +11,7 @@ class ColorController extends Controller
 {
     public function index() {
         if(session('admin_email')) {
+
         $siteInfo = Settings::all();
         $colors = Colors::all();
         return view('admin.colorsparameters.index', compact('siteInfo', 'colors'));
@@ -20,23 +21,35 @@ class ColorController extends Controller
     }
 
     public function delete($id) {
-        $deleted = Colors::destroy($id);
-        if($deleted) {
-            return redirect()->route('admin.colors.index')->with('success', 'Rəng uğurla silindi');
+        if(session('admin_email')) {
+            $deleted = Colors::destroy($id);
+            if($deleted) {
+                return redirect()->route('admin.colors.index')->with('success', 'Rəng uğurla silindi');
+            } else {
+                return redirect()->route('admin.colors.index')->with('error', 'Rəngin silinməsi zamanı xəta');
+            }
         } else {
-            return redirect()->route('admin.colors.index')->with('error', 'Rəngin silinməsi zamanı xəta');
+            return redirect()->route('admin.loginshow');
         }
     }
 
     public function store(Request $req) {
-        $all = $req->all();
-        $created = Colors::create($all);
-        if($created) {
+        if(session('admin_email')) {
+            $req->validate([
+                'color_name' => 'required',
+                'color' => 'required',
+            ]);
+            $all = $req->all();
+            $created = Colors::create($all);
             if($created) {
-                return redirect()->route('admin.colors.index')->with('success', 'Rəng uğurla əlavə edildi');
-            } else {
-                return redirect()->route('admin.colors.index')->with('error', 'Rəngin əlavə ediıməsi zamanı xəta');
+                if($created) {
+                    return redirect()->route('admin.colors.index')->with('success', 'Rəng uğurla əlavə edildi');
+                } else {
+                    return redirect()->route('admin.colors.index')->with('error', 'Rəngin əlavə ediıməsi zamanı xəta');
+                }
             }
+        } else {
+                return redirect()->route('admin.loginshow');
         }
     }
 }
